// app/stacks/components/landfills/ed-services/new-data-handler/src/apis/models/MonitorGeoSearch.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type MonitorGeoSearch = {
    /**
     * the latitude of the point to search
     */
    latitude: number;
    /**
     * the longitude of the point to search
     */
    longitude: number;
};